Move generateTestMods to its own file
