+++ src/redux/reducers.js
@@ -1,10 +1,8 @@
 import { createReducer } from "reduxsauce";
 import { Types } from "./actions";
-import { getSimplifiedModuleInfo } from "../utils/getSimplifiedModuleInfo";
+import { generateTestMods } from "../utils/generateTestMods";
 
 export const EMPTY_PLAN = { sems: {}, specialisations: [], majors: [], minors: [] };
-
-const simplifiedModules = getSimplifiedModuleInfo();
 
 const DEFAULT_SEMS = {
   Y1S1: { name: "Y1S1", mods: [] },
@@ -16,19 +14,6 @@ const DEFAULT_SEMS = {
   Y4S1: { name: "Y4S1", mods: [] },
   Y4S2: { name: "Y4S2", mods: [] },
 };
-
-// dummy mods for testing
-const generateTestMods = (sems) => {
-  let modIndex = 0;
-  Object.entries(sems).forEach(([key, semData]) => {
-    sems[key] = {
-      ...semData,
-      mods: Object.entries(simplifiedModules).slice(modIndex, modIndex + 5).map(([code, data]) => { return { code, ...data }; }),
-    };
-    modIndex += 5;
-  });
-  return sems;
-}
 
 export const INITIAL_STATE = {
   currentPlan: "sample plan",
